.agent/skills/generating-creative-content/requirements.txt
pyairtable>=2.2.0
requests>=2.31.0
google-generativeai>=0.4.0
python-dotenv>=1.0.1
